# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Clothing Line Commerce API
  description: |
    Headless e-commerce API for the clothing line storefront and admin consoles.

    This contract is the single source of truth shared by the storefront (Next.js), the
    backend adapters, and the mock (Prism). It is organised into six bounded contexts:

      - Admin     — catalogue administration, sales channels, regional pricing, tax,
                    promotions, roles, bulk imports, queue (dead-letter) operations.
      - Catalog   — storefront catalogue reads, search, category trees, cross-selling,
                    variant availability and product reviews.
      - Cart      — cart session lifecycle, line items (variant + custom B2B), discounts,
                    guest-cart merging.
      - Checkout  — shipping addresses, logistics quotes, insurance quotes, payment
                    session initialisation and async payment webhooks.
      - Customers — registration, JWT authentication, session revocation, password reset,
                    address books, B2B business units and quote requests.
      - Logistics — order returns (RMA), swaps, order edits, fulfilment dispatch, draft
                    orders and courier tracking webhooks.

    Conventions:

      * Money is always expressed as a non-negative **integer in minor units**
        (Kobo / Cents). Monetary fields end with the suffix `Minor` (e.g. `amountMinor`,
...
  version: 1.0.0
servers:
  - url: http://localhost:4000
...
security:
  - bearerAuth: []
paths:
...
  /store/payments/webhook:
    post:
      tags: [Checkout]
      summary: Payment gateway webhook (asynchronous payment confirmation).
      description: |
        Inbound payment-gateway webhook. The request body is consumed as RAW BYTES and
        verified against the HMAC-SHA512 signature in the `x-paystack-signature` header
        (computed over those exact bytes with the dedicated Paystack webhook secret — never
        the API secret key) BEFORE any JSON parsing; a mismatch returns `401` and the event
        is discarded.

        After verification the payload is validated, mapped to an internal payment event, and
        enqueued for background processing. The order is NEVER finalized synchronously from
        this request — the background PaymentEventWorker finalizes it later, idempotently by
        transaction reference. A `200` acknowledges receipt (and stops the gateway retrying);
        a `401` (invalid signature) or `500` (processing outage) causes the gateway to retry.
        Only `charge.success` events carrying our checkout metadata are processed; other
        events are acknowledged and ignored (`handled: false`).
      security: []
      parameters:
        - name: x-paystack-signature
          in: header
          required: true
          description: |
            HMAC-SHA512 (hex) signature of the raw request body, computed by Paystack with
            the dedicated webhook secret. Must be verified against the exact raw bytes.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaystackWebhookEvent"
      responses:
        "200":
          description: |
            Acknowledged. The verified event was enqueued for background processing
            (`handled: true`) or acknowledged but not applicable (`handled: false`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentWebhookAck"
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "401":
          $ref: "#/components/responses/StandardErrorResponse"
        "413":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"

  # =====================================================================
  # CUSTOMERS & AUTHENTICATION
  # =====================================================================
...
  /store/webhooks/courier-tracking:
    post:
      tags: [Logistics]
      summary: Courier tracking status webhook.
      description: |
        Accepts tracking events, updates the matching fulfilment, and optionally notifies the
        customer. Unknown tracking numbers are ignored idempotently; out-of-order events are
        dropped.

        SIGNATURE SECRET MECHANISM: the request body is consumed as RAW BYTES and verified
        against the HMAC-SHA512 signature in the `x-courier-signature` header (computed over
        those exact bytes with the dedicated `COURIER_TRACKING_WEBHOOK_SECRET` — never any API
        key) BEFORE any JSON parsing. A missing or mismatched signature returns `401` and the
        event is discarded (fail closed); the raw body is never parsed without a verified
        signature.

        After verification the payload is validated and mapped to a provider-neutral logistics
        event, then enqueued for background processing. This request NEVER updates fulfillment,
        NEVER creates shipments, NEVER calls the courier, and holds NO database transaction. A
        `204` acknowledges receipt (and stops the courier retrying); a `401` (invalid signature)
        or `500` (processing outage) causes the courier to retry; a `400` (malformed payload) is
        permanent and should not be retried.
      security: []
      parameters:
        - name: x-courier-signature
          in: header
          required: true
          description: |
            HMAC-SHA512 (hex) signature of the raw request body, computed by the courier with
            the dedicated webhook secret. Must be verified against the exact raw bytes.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourierTrackingWebhook"
      responses:
        "204":
          description: Tracking event processed (or ignored).
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "401":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /store/auth.
...
  responses:
    StandardErrorResponse:
      description: Standardized error envelope used by every non-2xx response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StandardError"
  schemas:
...
    StandardError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                # Mirrors the ErrorCode union in DomainError.ts exactly (one
                # entry per code, no duplicates, no stale codes). Keep this list
                # in sync with the domain union + mapDomainErrorToHttp.
                - VALIDATION_ERROR
                - INVALID_EMAIL
                - NEGATIVE_AMOUNT
                - INVALID_CURRENCY
                - INVALID_OPERATION
                - INVALID_STATE
                - INVALID_STATUS_TRANSITION
                - OUT_OF_STOCK
                - REGIONAL_PRICE_MISSING
                - INTERNAL_ERROR
                - JOB_PROCESSING_ERROR
                - PAYMENT_VERIFICATION_FAILED
                - LOGISTICS_VERIFICATION_FAILED
                - LOGISTICS_EVENT_FULFILLMENT_NOT_FOUND
                - EXTERNAL_SERVICE_TIMEOUT
                - EXTERNAL_SERVICE_UNAVAILABLE
                - EXTERNAL_SERVICE_ERROR
                - LOCK_ACQUISITION_FAILED
                - UNSUPPORTED_OPERATION
                - ORDER_ALREADY_FULFILLED
                - INVALID_RETURN_QUANTITY
                - DUPLICATE_DRAFT_ORDER
                - PAYMENT_REQUIRED
                - INVALID_RETURN_ITEM
                - INVALID_INPUT
                - REGION_NOT_FOUND
                - DUPLICATE_TRANSACTION
                - TRANSACTION_NOT_FOUND
                - INVALID_PAYMENT_AMOUNT
                - PERMISSION_DENIED
                - DUPLICATE_QUOTE
                - UNAUTHORIZED
                - UNAUTHORIZED_REVIEW
                - INVALID_CREDENTIALS
                - UNAUTHORIZED_ACCESS
                - CUSTOMER_ALREADY_EXISTS
                - COMPLIANCE_VIOLATION
                - ACCOUNT_DISABLED
                - ACCOUNT_LOCKED
                - BUSINESS_UNIT_ALREADY_EXISTS
                - PAYMENT_DECLINED
                - REFUND_REQUIRES_REVIEW
                - SHIPMENT_REQUIRES_RECONCILIATION
                - INVALID_SIGNATURE
                - RESOURCE_NOT_FOUND
                - CART_NOT_FOUND
                - PRODUCT_NOT_FOUND
                - INSUFFICIENT_INVENTORY
                - INSUFFICIENT_SINGLE_LOCATION_STOCK
                - SOURCING_FAILED
              example: OUT_OF_STOCK
            message:
              type: string
              example: Requested quantity exceeds available inventory.
            details:
              type: object
              additionalProperties: true

    # ------------------------------------------------------------------
    # Catalog DTOs
    # ------------------------------------------------------------------
...
    PaystackWebhookEvent:
      type: object
      required: [event, data]
      description: |
        The raw payment-gateway webhook payload as received by /store/payments/webhook.
        This is the PROVIDER's event shape (not our internal contract): the API verifies
        the signature over the raw bytes, validates this payload, and maps it to an
        internal payment event before enqueueing.
      properties:
        event:
          type: string
          description: Provider event type; only `charge.success` is processed.
          example: charge.success
        data:
          type: object
          required: [reference, amount]
          description: Provider event data.
          properties:
            reference:
              type: string
              description: >
                Provider transaction reference; echoed from the reference the API supplied
                at initialization. Used as the idempotency key for background finalization.
            amount:
              type: integer
              minimum: 0
              description: Amount captured by the provider in integer minor units (kobo/cents).
            currency:
              type: string
              description: >
                ISO-4217 currency code of the captured charge. The API verifies it against
                the durable payment obligation's currency; a mismatch is rejected.
            status:
              type: string
              description: Provider transaction status (e.g. success).
            metadata:
              type: object
              description: Metadata echoed from the initialization call.
              properties:
                cartId:
                  type: string
                  format: uuid
                  description: Checkout cart id; present only for checkout payments.
    PaymentWebhookAck:
      type: object
      required: [status, handled]
      properties:
        status:
          type: string
          const: ok
        handled:
          type: boolean
          description: >
            True when the verified event was accepted for background processing; false when
            it was acknowledged but not applicable (unrelated event type or no checkout
            metadata).
...
    CourierTrackingWebhook:
      type: object
      required: [trackingNumber, courierStatus, timestamp]
      properties:
        trackingNumber:
          type: string
        courierStatus:
          type: string
          enum: [in_transit, out_for_delivery, delivered, failed_attempt]
          description: Raw courier status; unknown values pass through.
        timestamp:
          type: string
          format: date-time
        notifyCustomer:
          type: boolean
          default: true
...